pyproject.toml
[project]
name = "imagetrainingsummaryquizgen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[all]>=0.116.1",
    "google-genai>=1.30.0",
    "ipython>=9.4.0",
    "jinja2>=3.1.6",
    "pydantic>=2.11.7",
    "pydotenv>=0.0.7",
    "youtube-transcript-api>=1.2.2",
]
